ci: Explicit "--target" for tests
When we're testing a target, Cargo needs to be instructed to compile
for that target, not for the host environment it's running in.

Otherwise, all these "arm-android" tests are just building for
"ubuntu-latest" host, which does not really test what it should test.
Once you do that, basically everything that is not compiling for the
host environment fails to run. Leave some FIXMEs there, for now CI
is going to be broken.