include all BoringSSL sources when publishing
Currently, `boring-sys` uses a Cargo [`include` field] to control what
files in the `boringssl` and `boringssl-fips` subdirectories are
actually published to crates.io. This means that only the specified
files are included in published versions of the crate.

Unfortunately, the FIPS version currently is missing a number of files
that it needs to build successfully. See issue #157 for details on this.

Complicating things, the Cargo.toml `include` field _only_ controls what
files are actually published to crates.io --- builds in this repository,
or using a Git dependency, will still include all files. This means that
there isn't an easy way to verify that all of the files actually
required to build BoringSSL are included without publishing a new
version of `boring-sys` to crates.io and checking that dependents can
build successfully. Therefore, I've changed the `include` directive to
simply include the entire BoringSSL source tree, rather than only
including some files in it. This does increase the size of the Cargo
package, but it has the substantial advantage that if the revisions of
the BoringSSL repo are changed, the new version of `boring-sys` won't be
silently broken, because it will still include all files that were in
the repo when we built it on CI. I discussed why this is necessary in
this comment:
https://github.com/cloudflare/boring/issues/157#issuecomment-1692274256

If we're really concerned about the size of the crates.io package, we
may want to instead *exclude* specific large test data files that aren't
actually required to build BoringSSL, instead of only *including* the
files we know are necessary. That way, any new files updated in
a future BoringSSL revision, or moved around, will still be included by
default, but we can avoid publishing large test vectors and other data
that we don't need to publish to crates.io.

Fixes #157

[`include` field]: https://doc.rust-lang.org/cargo/reference/manifest.html#the-exclude-and-include-fields