Specifically exclude large test data files
Rather than *including* only the files we believe are necessary to build
BoringSSL, this change adds an `exclude` section to the Cargo.toml to
explicitly exclude large test data files from the crates.io package.
I've verified that all the files excluded in this manner are okay to
exclude by manually deleting them from my local checkout and testing
that the build still works.

We may want to add automation for testing that this exclude
configuration is still okay when updating the BoringSSL revision, by
adding a shell script or something that deletes these files and tests
that we can build BoringSSL. This could run as part of the repo's CI
configuration.