boring: fix `clippy::unnecessary_cast` warnings
Updating Rust to 1.72.0 results in a bunch of new Clippy
[`unnecessary_cast`][1] warnings in the `boring` crate. This commit
fixes these warnings.

For the most part, the warnings are fixed by adding `#[allow]`
attributes for this lint. This is because many of these casts are
casting from `*mut c_uchar`/`*const c_uchar` to `*mut u8`/*const u8`.
Because the Clippy lint doesn't seem to know about type aliases, it sees
these as the same type. However, we don't necessarily know that
`c_uchar` will always be the same as `u8`, so in many instances, I've
allowed these warnings with a comment, rather than removing the cast
entirely.

[1]:
    https://rust-lang.github.io/rust-clippy/master/index.html#/unnecessary_cast