boring: allow `clippy::bad_bit_mask` lints
Updating to Rust 1.72.0 results in a bunch of new Clippy
[`bad_bit_mask`][1] lints showing up on uses of the `bitflags!` macro.

Clippy emits these warnings because it doesn't doesn't like `&` masking
with 0. This makes sense, because a mask of 0 always clears the entire
value, so the mask expression could just be the literal `0`. However, `&
0` shows up in the code generated by the `bitflags!` macro ---
presumably, this is because it's simpler to have the macro generate
expressions like `bits & 1 << 2`, `bits & 1 << 1`, and finally `bits & 1
<< 0` or similar, rather than special-casing the 0 case.

Thus, this commit allows this lint in modules that contain
`bitflags!`-generated code. Sadly, we can't just put the `#[allow]`
attribute on the bitflags types themselves, because the `bitflags!`
macro doesn't accept arbitrary metadata. So, we have to allow the lint
in the entire module when there's a `bitflags!` type present.

[1]:
    https://rust-lang.github.io/rust-clippy/master/index.html#/bad_bit_mask